phantom-trace-main/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.9.0
langchain>=1.2.0,<1.3.0
langchain-core>=1.2.0,<1.3.0
langchain-google-genai>=2.0.0
langgraph>=1.1.0,<1.2.0
langgraph-prebuilt>=1.0.9,<1.1.0
aiohttp>=3.10.0
python-dotenv==1.0.0